Now uses boost's split(). vector printer moved to its own header.
